% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servolabConnection.R
\name{ServolabGetConnection}
\alias{ServolabGetConnection}
\title{Estabilish a connection to Servolab Database}
\usage{
ServolabGetConnection(servolabIPaddress,username,password)
}
\arguments{
\item{servolabIPaddress}{the ipAddress of the server}

\item{username}{database username}

\item{password}{database password}
}
\value{
returns  a jdbc connection to Servolab
}
\description{
The function  connects to sybase using a JDBC driver. It takes in the host IP address
}
\examples{
host <-"172.11.11.11"
us <-"servo"
pass <- "pwd"
con <- ServolabGetConnection(host,us,pass)
}

