% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servolabGetResultsByMethodID.R
\name{ServolabGetResultsByMethodID}
\alias{ServolabGetResultsByMethodID}
\title{Get Sample Results from Servolab}
\usage{
ServolabGetResultsByMethodID( methodID, startDate, endDate, servoConnection)
}
\arguments{
\item{methodID}{a vector (integer) containing the identification number for a certain lab technique,  if you need all results just  create an ampty vector like c("")}

\item{startDate}{Date reception start}

\item{endDate}{Date receptioin end}
}
\value{
a data frame with lab results
}
\description{
The function takes in a study subject patern and creates a SQL statement to query the database and retrieve all records with the patientID matching the patern
}
\examples{
# First create a connection to servolab
con <- ServolabGetConnection("xxx.xxx.xxx.xxx",username,password)
startDate = '05/22/2016'
endDate = '07/22/2016'
method <- c(3212,3245)
result <- ServolabGetResultsByMethodID(method,startDate,endDate,con)
}

