% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servolabGetResultsByNidas.R
\name{ServolabGetResultsByNidas}
\alias{ServolabGetResultsByNidas}
\title{Get Sample Results from Servolab}
\usage{
ServolabGetResultsByNidas(nidasVector, methodID,servoConnection)
}
\arguments{
\item{nidasVector}{an vector containing NIDAS}

\item{methodID}{a vector (integer) which contains identification number for a certain lab technique,  if you need all results just create an ampty vector like c("")}

\item{servoConnection}{a JDBCConnection  to servolab}
}
\value{
returns a data frame with lab results
}
\description{
The function takes in an vector of Nidas and creates a SQL statement to query the database
}
\examples{
# First create a connection to servolab
con <- ServolabGetConnection("xxx.xxx.xxx.xxx",username,password)
# Then get the results
method_id   <- c(3212,3245)
result <- ServolabGetResultsByNidas(c('1234567.8','15555567.5'),method_id,con)
}

