% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servolabGetResultsBySubjectID.R
\name{ServolabGetResultsBySubjectID}
\alias{ServolabGetResultsBySubjectID}
\title{Get Sample Results from Servolab}
\usage{
ServolabGetResultsBySubjectID(subjectID, methodID, year,servoConnection)
}
\arguments{
\item{subjectID}{the subjectID pattern ex: "GAMA" ,"XMAL","VIMAG", etc.}

\item{methodID}{a vector (integer) containing the identification number for a certain lab technique,  if you need all results ust create an ampty vector like c("")}

\item{year}{the year when the samples were processed, if you dont want to specify , leave it empty: ""}
}
\value{
a data frame with lab results
}
\description{
The function takes in a study subject patern and creates a SQL statement to query the database and retrieve all records with the patientID matching the patern
}
\examples{
# First create a connection to servolab
con <- ServolabGetConnection("xxx.xxx.xxx.xxx",username,password)
# Then get the results
patern <- "INPD"
year <- 2014
method <- c(3212,3245)
result <- ServolabGetResultsBySubjectID(patern,method,year,con)
}

